fix: Refactor inspector.js into ES modules
